asm fini

+++ asm/srcs/add_word.c
@@ -6,7 +6,7 @@
 /*   By: brvalcas <brvalcas@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/06/17 15:39:32 by bryanvalcas       #+#    #+#             */
-/*   Updated: 2019/07/05 15:43:28 by brvalcas         ###   ########.fr       */
+/*   Updated: 2019/07/06 13:12:56 by brvalcas         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -46,7 +46,7 @@ void			add_word(t_data *data, t_token word)
 	{
 		j = i;
 		check_char(word, &i, &j);
-		if (!(new.cut = ft_strcut(word.cut, i, j)))
+		if (!(new.cut = ft_strcut(word.cut, (size_t)i, (size_t)j)))
 			return ;
 		add_token(&data->token, new);
 		i = (j - 1 >= 0) ? j - 1 : j;
