+++ asm/srcs/list_chain_new.c
@@ -6,7 +6,7 @@
 /*   By: brvalcas <brvalcas@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/06/16 22:57:45 by bryanvalcas       #+#    #+#             */
-/*   Updated: 2019/07/05 15:48:49 by brvalcas         ###   ########.fr       */
+/*   Updated: 2019/07/06 13:08:17 by brvalcas         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -53,7 +53,7 @@ t_name_label	*new_n_label(char *label, int index)
 	if (!(new = malloc(sizeof(t_ins))))
 		return (NULL);
 	new->label = (label) ? label : NULL;
-	new->len = (label) ? ft_strlen(label) : 0;
+	new->len = (label) ? (int)ft_strlen(label) : 0;
 	new->index_ins = index;
 	new->next = NULL;
 	return (new);
